First sketch
Started with the idea of a tower-defense game, might get more
complicated, for now focused on castle construction issues.

Recycled the HexagonPatch code, wrote code to extend a castle from the
center of a patch by adding rings around known geometry. Once the a ring
touches the border we go a step back to have some breathing room and no
walls directly touching a patch border.

Then we reserve two rings for the wall, and use the inner rings to
construct our village. We always consider pairs of rings because
topologically that gives us rings where we don't have to move diagonal
to stay within the ring.

We alternate for ring pairs: Either DIRT/free or a 66% of building plot
and some lesser chance of actual building. Add some greenery.

Currently struggling with drawing an optimized castle wall on the 2
rings reserved for the wall.